Finer object layering
Game objects are now split into separate lists: stage objects,
background spawned objects, fighter objects, and foreground spawned
objects. This essentially forces objects to be in particular "layers"
for the sake of rendering.

Also, the fighter objects are re-arranged every time fighters strike one
another. The attacker is moved to be drawn on top of the other fighter
objects.