More GUI upgrades
Placeholder graphics and text are now defineable in the playable
characters list definition file. All are optional.

Instructions text shown during character select is now editable.

HSFont definition now has a Palettes section, just like HSObjects. Fonts
listed there will be applied to every character in the font. Anywhere a
font is specified in a definition, a palette index can also be
specified.

Fonts are not required to be defined. This is mostly for the sake of
having menus with no title/header.

Health meter cover, combo counter background, and prompt background are
now optional in the match hud definition. Also, the combo counter text
is now editable.

Justifications can be set for menu titles and items.