Defensive attribute changes
Altered Blockability to only be Blockable or Unblockable. Created
HitLevel to determine if an attack is high, low, or mid. Added "strictly
mid" to signify attacks that are ALWAYS mid and ignore vertical velocity
and position.

Changed backslashes to forward slashes in hard coded file paths.

Made hitstop a universal value, but allow attacks to override it.

Added super armor.